pyproject.toml
[project]
name = "homedns"
dynamic = ["version"]  # single source the projects version
requires-python = ">= 3.10"
dependencies = [
    "pyopenssl >= 24.1.0",
    "twisted >= 24.3.0",
    "service_identity",
    "klein >= 23.12.0",
    "pyjwt >= 2.8.0",
    "pyyaml >= 6.0.1",
    "tabulate",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools.dynamic]
version = {attr = "homedns.__version__"}

[project.urls]
Homepage = "https://github.com/bendemott/homedns"
Documentation = "https://github.com/bendemott/homedns"
Repository = "https://github.com/bendemott/homedns"
Issues = "https://github.com/bendemott/homedns/issues"

[project.scripts]
homedns-server = "homedns.cli.server:main"
homedns-api = "homedns.cli.restclient:main"
homedns-jwt = "homedns.cli.jwt:main"